Optimization 2: Don't copy input row for rotation of 63 bits.
This rotation uses a gate that constrains the rotation result with the row that is immediately above. But rotations are always called using the last row of the trace as operand. In this commit we take advantage on that and just write the result instead of copying, without changing the gate