Optimization 3: Addition copying just one operand
In the mix function, all additions uses the last row of the trace as one of the parameters. This lets us use that row and just copy the second parameter of the addition.

Since this does not happen outside the mix (or in the first addition in the mix) we have two addition methods and just call the optimized version when the conditions are met